Test and restructure Map and Player Class

+++ ProiectMC++/ProiectMC++/GameObject.cpp
@@ -1,16 +1,13 @@
 #include "GameObject.h"
 
-std::pair<int16_t, int16_t> GameObject::getPosition() const
-{
+std::pair<size_t, size_t> GameObject::getPosition() const{
 	return m_position;
 }
 
-void GameObject::setPostion(std::pair<int16_t, int16_t> t_position)
-{
+void GameObject::setPostion(std::pair<int16_t, int16_t> t_position){
 	m_position = t_position;
 }
 
-void GameObject::setSpeed(int16_t t_speed)
-{
+void GameObject::setSpeed(int16_t t_speed){
 	m_speed = t_speed;
 }
